\docType{package}
\name{BET_package}
\alias{BET-package}
\title{Binary Expansion Testing}
\description{
The \code{BET} package provides functions for nonparametric detection of nonuniformity and dependence with Binary Expansion Testing (BET).
}
\section{BET functions}{

\code{\link{MaxBET}}
\code{\link{symm}}
\code{\link{get.signs}}
\code{\link{cell.counts}}
\code{\link{bet.plot}}
\code{\link{MaxBETs}}
\code{\link{BEAST}}
}

\section{Reference(s)}{

Kai Zhang (2019) BET on Independence, Journal of the American Statistical Association, 114:528, 1620-1637, \doi{10.1080/01621459.2018.1537921},
Kai Zhang, Zhigen Zhao, and Wen Zhou (2021). BEAUTY Powered BEAST, <arXiv:2103.00674> and
Wan Zhang, Zhigen Zhao, Michael Baiocchi, Yao Li, Kai Zhang. SorBET: A Fast and Powerful Algorithm to Test Dependence of Variables. Techinical report, 2023.
}

