\docType{package}
\name{BET_package}
\alias{BET_package}
\title{Binary Expansion Testing}
\description{
The \code{BET} package provides functions for nonparametric detection of nonuniformity and dependence with Binary Expansion Testing (BET).
}
\section{BET functions}{

\code{\link{BET}}
\code{\link{symm}}
\code{\link{bet.plot}}
\code{\link{BETs}}
}

\section{Reference(s)}{

Kai Zhang (2019) BET on Independence, Journal of the American Statistical Association, 114:528, 1620-1637, <DOI:\href{https://doi.org/10.1080/01621459.2018.1537921}{10.1080/01621459.2018.1537921}> and
Zhigen Zhao, Michael Baiocchi, Kai Zhang. SorBET: A Fast and Powerful Algorithm to Test Dependence of Variables.
}

